+++ src/r_draw.c
@@ -34,6 +34,8 @@
 #include "v_video.h" // Needs access to LFB (guess what).
 #include "doomstat.h" // State.
 
+#include <stdint.h>
+
 
 // ?
 #define MAXWIDTH 1120
@@ -371,7 +373,7 @@ void R_InitTranslationTables(void)
     int i;
 
     translationtables = Z_Malloc(256 * 3 + 255, PU_STATIC, 0);
-    translationtables = (byte*)(((unsigned long long)translationtables + 255) & ~255);
+    translationtables = (byte*)(((uintptr_t)translationtables + 255) & ~255);
 
     // translate just the 16 green colors
     for (i = 0; i < 256; i++)
